+++ NEWS.md
@@ -1,7 +1,11 @@
-# bestNormalize 0.2.2.9001
+# bestNormalize 0.2.2.9002
+
+- Added experimental feature to estimate out-of-sample normality statistics in bestNormalize 
+  instead of in-sample ones via repeated cross-validation
+    - Note: set `out_of_sample = FALSE` to maintain backward-compatibility with prior versions
 
 - Improved extrapolation of the ORQ (orderNorm) method
-    - Instead of linear extrapolation, it uses binomial (logit-link) model on ranks,
+    - Instead of linear extrapolation, it uses binomial (logit-link) model on ranks
     
 - Added plotting feature for transformation objects